Extracting out the correct order to a common place
