+++ Specifications/CodeAnalysis/ElementsMustAppearInTheCorrectOrder/Constructors/UnitTests.cs
@@ -1,42 +1,11 @@
 namespace Aksio.CodeAnalysis.ElementsMustAppearInTheCorrectOrder.Constructors
 {
-    public class UnitTests: CodeFixVerifier
+    public class UnitTests : CodeFixVerifier
     {
         [Fact]
         public void CorrectOrder()
         {
-            const string content = @"
-                class Blabla
-                {
-                    string _someBacking;
-                    public string BackedField => _someBacking.Replace(""a"", ""b"");
-                    public int Teller { get; private set; }
-                    public delegate void SomethingHappenedEventHandler(object sender, object args);
-                    public event EventHandler SomethingHappened;
-
-                    public Blabla()
-                    {
-                        Teller = 0;
-                        _someBacking = ""meh"";
-                    }
-
-                    ~Blabla()
-                    {
-                        Teller = int.MinValue;
-                    }
-
-                    public int this[int i] => Teller;
-
-                    public void GjørNoeRart()
-                    {
-                        _someBacking = $""{ØkTeller()}"";
-                    }
-
-                    void ØkTeller() => ++Teller;
-                }   
-            ";
-
-            VerifyCSharpDiagnostic(content);
+            VerifyCSharpDiagnostic(Common.ValidOrder);
         }
 
         [Fact]
