+++ Specifications/CodeAnalysis/ElementsMustAppearInTheCorrectOrder/Fields/UnitTests.cs
@@ -1,42 +1,11 @@
 namespace Aksio.CodeAnalysis.ElementsMustAppearInTheCorrectOrder.Fields
 {
-    public class UnitTests: CodeFixVerifier
+    public class UnitTests : CodeFixVerifier
     {
         [Fact]
         public void FieldsBeforeConstructor()
         {
-            const string content = @"
-                class Blabla
-                {
-                    string _someBacking;
-                    public string BackedField => _someBacking.Replace(""a"", ""b"");
-                    public int Teller { get; private set; }
-                    public event EventHandler SomethingHappened;
-                    public delegate void SomethingHappenedEventHandler(object sender, object args);
-
-                    public Blabla()
-                    {
-                        Teller = 0;
-                        _someBacking = ""meh"";
-                    }
-
-                    ~Blabla()
-                    {
-                        Teller = int.MinValue;
-                    }
-
-                    public int this[int i] => Teller;
-
-                    public void GjørNoeRart()
-                    {
-                        _someBacking = $""{ ØkTeller()}
-                        "";
-                    }
-                    void ØkTeller() => ++Teller;
-                }
-            ";
-
-            VerifyCSharpDiagnostic(content);
+            VerifyCSharpDiagnostic(Common.ValidOrder);
         }
 
         [Fact]
