+++ Specifications/CodeAnalysis/ElementsMustAppearInTheCorrectOrder/Methods/UnitTests.cs
@@ -1,42 +1,11 @@
 namespace Aksio.CodeAnalysis.ElementsMustAppearInTheCorrectOrder.Methods
 {
-    public class UnitTests: CodeFixVerifier
-    { 
+    public class UnitTests : CodeFixVerifier
+    {
         [Fact]
         public void CorrectOrder()
         {
-            const string content = @"
-                class Blabla
-                {
-                    string _someBacking;
-                    public string BackedField => _someBacking.Replace(""a"", ""b"");
-                    public int Teller { get; private set; }
-                    public event EventHandler SomethingHappened;
-                    public delegate void SomethingHappenedEventHandler(object sender, object args);
-
-                    public Blabla()
-                    {
-                        Teller = 0;
-                        _someBacking = ""meh"";
-                    }
-
-                    ~Blabla()
-                    {
-                        Teller = int.MinValue;
-                    }
-
-                    public int this[int i] => Teller;
-
-                    public void GjørNoeRart()
-                    {
-                        _someBacking = $""{ØkTeller()}"";
-                    }
-
-                    void ØkTeller() => ++Teller;
-                }   
-            ";
-
-            VerifyCSharpDiagnostic(content);
+            VerifyCSharpDiagnostic(Common.ValidOrder);
         }
 
         [Fact]
@@ -68,7 +37,7 @@ namespace Aksio.CodeAnalysis.ElementsMustAppearInTheCorrectOrder.Methods
 
             VerifyCSharpDiagnostic(content, GetExpectedFailure());
         }
-        
+
         [Fact]
         public void MethodsBeforeDelegates()
         {
@@ -83,7 +52,7 @@ namespace Aksio.CodeAnalysis.ElementsMustAppearInTheCorrectOrder.Methods
 
             VerifyCSharpDiagnostic(content, GetExpectedFailure());
         }
-        
+
         [Fact]
         public void MethodsBeforeEvents()
         {
