+++ Specifications/CodeAnalysis/ElementsMustAppearInTheCorrectOrder/Properties/UnitTests.cs
@@ -5,41 +5,7 @@ namespace Aksio.CodeAnalysis.ElementsMustAppearInTheCorrectOrder.Properties
         [Fact]
         public void CorrectOrder()
         {
-            const string content = @"
-                class Blabla
-                {
-                    string _someBacking;
-                    public string BackedField => _someBacking.Replace(""a"", ""b"");
-                    public int Teller { get; private set; }
-                    public event EventHandler SomethingHappened;
-                    public delegate void SomethingHappenedEventHandler(object sender, object args);
-
-                    public Blabla()
-                    {
-                        Teller = 0;
-                        _someBacking = ""meh"";
-                    }
-
-                    ~Blabla()
-                    {
-                        Teller = int.MinValue;
-                    }
-
-                    public int this[int i] => Teller;
-
-                    public void GjørNoeRart()
-                    {
-                        _someBacking = $""{ ØkTeller()}
-                        "";
-                    }
-                    void ØkTeller() => ++Teller;
-
-                    // Private properties should be ignored
-                    int TellerPlus1 => Teller+1;
-                }
-            ";
-
-            VerifyCSharpDiagnostic(content);
+            VerifyCSharpDiagnostic(Common.ValidOrder);
         }
 
         [Fact]
